Timber class: move to Twig namespace. Old location was confusing.
